refactor: separate events
