+++ packages/blocky-core/src/model/index.ts
@@ -1,6 +1,6 @@
-
 export { default as State, DocNodeName } from "./state";
 export { type CursorState } from "./cursor";
 export { TextType } from "./text";
 export * from "./tree";
 export * from "./element";
+export * from "./events";
